requirements.txt
tweepy==3.8.0
speedtest-cli==2.1.2